+++ packages/integration-sdk/CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to
 ### Changed
 
 - [#321](https://github.com/JupiterOne/sdk/issues/321) -
-  `j1-integration document` output in alphabetical order by `_type`.
+  `j1-integration document` output in alphabetical order by entity metadata
+  `resourceName` property and relationship metadata `_type` property.
 
 ## 3.1.0 - 2020-08-26